Image fetch race condition fixed
There was a reoccurring error (every fifth or so time) where
canvas.NewImageFromURI() would panic with a found free pointer
(during http.(*Transport).GetConn). This was despite not having
any concurrency myself. The issue probably stems from
Fyne's internal structure.

Instead of fetching the image right away, a placeholder is inserted and
the image is fetched after the whole tree is rendered and inserted into
the window. Then its container is refreshed.